requirements.txt
requests
torch
readability
bitsandbytes
accelerate
pandas
datasets
transformers[torch]>=4.40.0
accelerate>=0.28.0
numpy<2.0
evaluate
scikit-learn
tqdm
huggingface-hub
hf_xet  # For faster Xet Storage downloads from HuggingFace Hub
Pillow
nltk
rouge
rouge-score
absl-py
codecarbon
setuptools
wandb
mysql-connector-python
peft>=0.7.0  # For LoRA and parameter-efficient fine-tuning
bert-score  # For dataset quality analysis

# for the converter (updated for Python 3.13 compatibility)
onnxruntime>=1.20.0
optimum>=1.21.0,<2.0.0
onnx>=1.16.0,<1.20  # Pin to <1.20 for ONNXRuntime quantization compatibility
onnxscript>=0.5.0  # Required by torch.onnx.export
